fix: changed data type from int to size_t

+++ inc/ConfigFileParser.hpp
@@ -44,8 +44,8 @@ private:
 	size_t m_serverIndex;
 	std::vector<std::string> m_validServerDirectives;
 	size_t m_locationIndex;
-	int m_serverRootCount;
-	int m_locationRootCount;
+	size_t m_serverRootCount;
+	size_t m_locationRootCount;
 	std::vector<std::string> m_validLocationDirectives;
 	bool m_isDefaultLocationDefined;
 	static const char* const s_whitespace;
